# docs/swagger.yaml
basePath: /
definitions:
  domain.AuthResponse:
    properties:
      accesstoken:
        type: string
      refreshtoken:
        type: string
      user:
        $ref: '#/definitions/domain.User'
    type: object
  domain.Permissions:
    properties:
      product_module:
        type: boolean
      profile_module:
        type: boolean
      user_module:
        type: boolean
    type: object
  domain.Product:
    properties:
      id:
        type: integer
      name:
        minLength: 2
        type: string
    required:
    - name
    type: object
  domain.Profile:
    properties:
      id:
        type: integer
      name:
        minLength: 4
        type: string
      permissions:
        $ref: '#/definitions/domain.Permissions'
    required:
    - name
    type: object
  domain.User:
    properties:
      id:
        type: integer
      mail:
        type: string
      name:
        minLength: 5
        type: string
      new:
        type: boolean
      profile:
        $ref: '#/definitions/domain.Profile'
      status:
        type: boolean
    required:
    - mail
    - name
    type: object
  dto.ItemsOutputDTO:
    properties:
      count:
        type: integer
      items: {}
    type: object
  dto.LoginInputDTO:
    properties:
      email:
        example: admin@admin.com
        type: string
      password:
        example: "12345678"
        type: string
    type: object
...
  dto.ProductInputDTO:
    properties:
      name:
        example: Categoria 01
        type: string
    type: object
...
  httphelper.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
info:
  contact:
    email: email@email.com
    name: Raul del Aguila
  description: Expenses API.
  title: Go - Expenses API
paths:
...
  /auth:
    get:
      consumes:
      - application/json
      description: User authenticated
      parameters:
      - description: User token
        in: header
        name: Authorization
        type: string
      - description: Language responses
        in: query
        name: lang
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.User'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
      security:
      - Bearer: []
      summary: User authenticated
      tags:
      - Auth
    post:
      consumes:
      - application/json
      description: User authentication
      parameters:
      - description: Language responses
        in: query
        name: lang
        type: string
      - description: Credentials model
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dto.LoginInputDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.AuthResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
      summary: User authentication
      tags:
      - Auth
    put:
      consumes:
      - application/json
      description: User refresh
      parameters:
      - description: User token
        in: header
        name: Authorization
        type: string
      - description: Language responses
        in: query
        name: lang
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.AuthResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
      summary: User refresh
      tags:
      - Auth
  /product:
    get:
      consumes:
      - application/json
      description: Get categories
      parameters:
      - description: Language responses
        in: query
        name: lang
        type: string
      - example: 10
        in: query
        name: limit
        type: integer
      - example: descending order 'desc' or ascending order 'asc'
        in: query
        name: order
        type: string
      - example: 1
        in: query
        name: page
        type: integer
      - example: name
        in: query
        name: search
        type: string
      - example: '''updated_at'', ''created_at'', ''name'' or some other field of
          the response object'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ItemsOutputDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
      security:
      - Bearer: []
      summary: Get categories
      tags:
      - Product
    post:
      consumes:
      - application/json
      description: Insert product
      parameters:
      - description: Language responses
        in: query
        name: lang
        type: string
      - description: Product model
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/dto.ProductInputDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
      security:
      - Bearer: []
      summary: Insert product
      tags:
      - Product
  /product/{id}:
    delete:
      consumes:
      - application/json
      description: Delete product by ID
      parameters:
      - description: Language responses
        in: query
        name: lang
        type: string
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
      security:
      - Bearer: []
      summary: Delete product by ID
      tags:
      - Product
    get:
      consumes:
      - application/json
      description: Get product by ID
      parameters:
      - description: Language responses
        in: query
        name: lang
        type: string
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
      security:
      - Bearer: []
      summary: Get product by ID
      tags:
      - Product
    put:
      consumes:
      - application/json
      description: Update product by ID
      parameters:
      - description: Language responses
        in: query
        name: lang
        type: string
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product model
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/dto.ProductInputDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
      security:
      - Bearer: []
      summary: Update product by ID
      tags:
      - Product
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"